requirements.txt
google-cloud-speech>=0.34.0
pyaudio>=0.2.11
boto3
awscli
pygame
playsound
pydispatch
